test: add test utils file
